inital commit

+++ src/Daicode/Bai03.java
@@ -6,7 +6,7 @@ public class Bai03 {
 	public static void main(String[] args) {
 		Scanner scanner = new Scanner(System.in);
 		System.out.println("canh :");
-		double canh = scanner.nextInt();
+		int canh = scanner.nextInt();
 		System.out.println("the tich = " + Math.pow(canh, 3));
 		scanner.close();
 	}
